pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "vcon-mac-wtf"
version = "0.1.0"
description = "MLX Whisper transcription server for Apple Silicon with WTF/vCon support"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "mlx-whisper>=0.4.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "python-multipart>=0.0.9",
    "python-dotenv>=1.0.0",
    "vcon-wtf>=0.1.1",
]

[project.urls]
Homepage = "https://github.com/vcon-dev/vcon-mac-wtf"
Repository = "https://github.com/vcon-dev/vcon-mac-wtf"
Documentation = "https://github.com/vcon-dev/vcon-mac-wtf#readme"
"Bug Tracker" = "https://github.com/vcon-dev/vcon-mac-wtf/issues"

[project.scripts]
vcon-mac-wtf = "vcon_mac_wtf.main:run"

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",
    "ruff>=0.5.0",
    "mypy>=1.10.0",
]

[tool.hatch.build.targets.wheel]
packages = ["src/vcon_mac_wtf"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "integration: requires Apple Silicon and MLX (deselect with '-m not integration')",
]

[tool.ruff]
line-length = 99
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W"]
